fix llm.ts
